Simplify recycler config by using a supervol UUID
This simplifies the config of the recycler pod by using a UUID file in
the supervol to determine which PVs it should be recycling. The simpler
config lessens the chance of misconfiguration that leads to either no
PVs getting recycled or accidently recycling the wrong ones.
This also adds locking to the UUID file to prevent multiple recyclers
from running against the same supervol.

Signed-off-by: John Strunk <jstrunk@redhat.com>