Send logs to container stdout
This adds a new container in the flex DS, "logs" that runs rsyslog to
send all gluster log files to the container's stdout.
There is also a logrotate container (currently inactive). The logrotate
container is not currently used because all log files are going into the
host's /var/log/* tree and getting rotated by the host's rotation
mechanisms.

- Flex plugin log has been changed to /var/log/glusterfs/glfs-subdir.log
(instead of /tmp/glfs-subdir.out).
- Fuse (glusterfs) supervol logfile is now set manually to be
/var/log/glusterfs/<serverip>-<volume>.log to shorten the filename and
make it easier to locate a particular supervol's log.

Signed-off-by: John Strunk <jstrunk@redhat.com>