Adds a helper script to the recycler pod to use kubectl
Realizing the recycler has the permissions to modify PVs makes it a
reasonable place to add the static PVs to the cluster. The helper script
sets the kubectl command options to use the service account.

One can now /kubectl-sa.sh apply -f /path/to/pvs.yml to add them to the
cluster. This works out well since the creation script dumps them into
the supervol root that the recycler already mounts!

Signed-off-by: John Strunk <jstrunk@redhat.com>